requirements.txt
azure-cosmos
peewee
requests
